refactor: view용 Rank, Suit enum 추가
